+++ plugins/login-ip.php
@@ -11,26 +11,31 @@ class AdminerLoginIp extends Adminer\Plugin {
 
 	/** Set allowed IP addresses
 	* @param list<string> $ips IP address prefixes
-	* @param list<string> $forwarded_for X-Forwarded-For prefixes if IP address matches, empty array means anything
+	* @param list<string> $forwarded_for X-Forwarded-For prefixes if IP address matches, empty array means that the request must not be proxied
 	*/
-	function __construct(array $ips, array $forwarded_for = array()) {
+	function __construct(array $ips = array('127.', '::1'), array $forwarded_for = array()) {
 		$this->ips = $ips;
-		$this->forwarded_for= $forwarded_for;
+		$this->forwarded_for = $forwarded_for;
 	}
 
 	function login($login, $password) {
-		foreach ($this->ips as $ip) {
-			if (strncasecmp($_SERVER["REMOTE_ADDR"], $ip, strlen($ip)) == 0) {
-				if (!$this->forwarded_for) {
-					return true;
-				}
-				if ($_SERVER["HTTP_X_FORWARDED_FOR"]) {
-					foreach ($this->forwarded_for as $forwarded_for) {
-						if (strncasecmp(preg_replace('~.*, *~', '', $_SERVER["HTTP_X_FORWARDED_FOR"]), $forwarded_for, strlen($forwarded_for)) == 0) {
-							return true;
-						}
-					}
-				}
+		// a proxy appends the client to X-Forwarded-For, the preceding values are sent by the client itself
+		$forwarded_for = preg_replace('~.*, *~', '', strval($_SERVER["HTTP_X_FORWARDED_FOR"]));
+		return $this->matchPrefix($_SERVER["REMOTE_ADDR"], $this->ips)
+			&& ($this->forwarded_for
+				? $this->matchPrefix($forwarded_for, $this->forwarded_for)
+				: $forwarded_for == "" // an unexpected proxy would pass requests from anywhere
+			);
+	}
+
+	/** Check if the value begins with one of the prefixes
+	* @param list<string> $prefixes
+	* @return bool
+	*/
+	private function matchPrefix($value, array $prefixes) {
+		foreach ($prefixes as $prefix) {
+			if (strncasecmp(strval($value), $prefix, strlen($prefix)) == 0) {
+				return true;
 			}
 		}
 		return false;